python-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generative-ai==0.3.0
firebase-admin==6.4.0
pymongo==4.6.1
psycopg2-binary==2.9.9
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
requests==2.31.0
pyjwt==2.8.0
cryptography==41.0.7

